+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: `StickyContainer`: Add `"never"` value for `isSticky` prop.
+- Feature: `HeaderMenu`: Improve accessibility.
+- Feature: `HeaderNav.UserMenuNext`: Add `mobilePadding` prop.
 
 ## [3.21.0] - 2021-06-21